cambio tipo de dato de cantidad

+++ src/main/java/daw2a/gestion_alimentos_api_rest/entities/Existencia.java
@@ -48,7 +48,7 @@ public class Existencia {
     @Positive(message = "La cantidad debe ser un número positivo")
     @NotNull(message = "La cantidad no puede ser nula")
     @NotBlank(message = "La cantidad no puede estar vacía")
-    private int cantidad;
+    private Long cantidad;
 
     /**
      * Fecha en que el alimento fue colocado en esa ubicación
